+++ backend/app/controllers/time_blocks_controller.rb
@@ -57,7 +57,7 @@ class TimeBlocksController < WebController
   def authorize_senior_access!
     # User must be the senior or a caregiver for the senior
     unless current_user == @senior || current_user.seniors.include?(@senior)
-      redirect_to dashboard_path, alert: "You don't have access to this senior's time blocks"
+      redirect_to dashboard_path, alert: "You don't have access to this care receiver's time blocks"
     end
   end